# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Epigra Assigment",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": ["Authentication"],
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": ["Authentication"],
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": ["Authentication"],
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/capsules": {
            "get": {
                "tags": ["Capsule"],
                "summary": "Get all capsules",
                "description": "Returns all capsules",
                "operationId": "getCapsules",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App/Models/Capsule.php"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/{status}/capsules": {
            "get": {
                "tags": ["Capsule"],
                "summary": "Get all capsules by status",
                "description": "Returns all capsules by status",
                "operationId": "getCapsulesByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Status of the capsule",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App/Models/Capsule.php"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/capsules/{serial}": {
            "get": {
                "tags": ["Capsule"],
                "summary": "Find capsule by serial",
                "description": "Returns a single capsule details",
                "operationId": "getCapsuleBySerial",
                "parameters": [
                    {
                        "name": "serial",
                        "in": "path",
                        "description": "Serial number of the capsule",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Models\\Capsule.php"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid serial number"
                    },
                    "404": {
                        "description": "Capsule not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}